refactor: rename 'tsv-file-reader.ts' from 'tsv.reader.ts'
